Moved SendPageForm to newly-created send-page folder under components
